api/requirements.txt
flask
inference-sdk
opencv-python
numpy
Pillow
python-dotenv
reportlab
pandas
